+++ ZeShmouttsAssets/Blizzard API for Unity/Scripts/WoW Game Data API/Mythic Keystone Leaderboard API/Endpoint_WoW_GameData_MythicKeystoneLeaderboardsIndex.cs
@@ -16,16 +16,17 @@ namespace ZeShmouttsAssets.BlizzardAPI
 		public static partial class WowGameData
 		{
 			/// <summary>
-			/// Coroutine that retrieves an index of all WoW Mythic Keystone seasons.
+			/// Coroutine that retrieves an index of WoW Mythic Keystone leaderboards for a connected realm.
 			/// </summary>
+			/// <param name="connectedRealmId">The ID of the connected realm.</param>
 			/// <param name="action_Result">Action to execute with the data once retrieved and converted.</param>
 			/// <param name="ifModifiedSince">Adds a request header to check if the document has been modified since this date (in HTML format), which will return an empty response body if it's older.</param>
 			/// <param name="action_LastModified">Action to execute with the date of the last server-side modification to the document.</param>
 			/// <param name="region">The region of the data to retrieve.</param>
 			/// <returns></returns>
-			public static IEnumerator GetMythicKeystoneSeasonsIndex(Action<WowMythicKeystoneSeasonsIndex_JSON> action_Result, string ifModifiedSince = null, Action<string> action_LastModified = null, BattleNetRegion region = BattleNetRegion.UnitedStates)
+			public static IEnumerator GetMythicKeystoneLeaderboardsIndex(int connectedRealmId, Action<WowMythicKeystoneLeaderboardsIndex_JSON> action_Result, string ifModifiedSince = null, Action<string> action_LastModified = null, BattleNetRegion region = BattleNetRegion.UnitedStates)
 			{
-				string path = "/data/wow/mythic-keystone/season/index";
+				string path = string.Format("/data/wow/connected-realm/{0}/mythic-leaderboard/index", connectedRealmId);
 				yield return SendRequest(region, namespaceDynamic, path, action_Result, ifModifiedSince, action_LastModified);
 			}
 		}
@@ -35,14 +36,13 @@ namespace ZeShmouttsAssets.BlizzardAPI
 namespace ZeShmouttsAssets.BlizzardAPI.JSON
 {
 	/// <summary>
-	/// JSON structure for an index of World of Warcraft Mythic Keystone seasons.
+	/// JSON structure for World of Warcraft Mythic Keystone leaderboards.
 	/// </summary>
 	[Serializable]
-	public class WowMythicKeystoneSeasonsIndex_JSON : Object_Json
+	public class WowMythicKeystoneLeaderboardsIndex_JSON : Object_Json
 	{
 		public LinkStruct _links;
 
-		public RefIdStruct[] seasons;
-		public RefIdStruct current_season;
+		public RefNameIdStruct[] current_leaderboards;
 	}
 }